+++ Balances.txt
@@ -1,4 +1,4 @@
-80.5 
+70.5 
 89193.0 
 99999945466.0 
 100.0
